api_server/requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
click==8.1.3
fastapi==0.75.2
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing_extensions==4.2.0
uvicorn==0.17.6
argon2-cffi
argon2-cffi-bindings
passlib
fastapi-jwt-auth
